Refactor GMRES to use existing Arnoldi
